compio: split send.rs into send/ directory module
Extract routing strategies from the 1023-line monolithic send.rs into
separate files for readability:

- send/mod.rs: dispatch + shared helpers (try_direct_encode, direct_push_*)
- send/round_robin.rs: PUSH/DEALER/REQ/PAIR/REP
- send/fan_out.rs: PUB/XPUB/XSUB (subscription-filtered)
- send/identity.rs: ROUTER/SERVER/PEER (identity-routed)
- send/radio.rs: RADIO (UDP + group-filtered ZMTP)

Purely organizational: all methods stay on Socket, no new traits or
enums, zero runtime change. Mirrors omq-tokio's routing/ split.

Also fix three pre-existing clippy warnings in pub_sub.rs test
(doc_markdown, needless_range_loop, manual_let_else).
